V11.2: 修复 DDP temporal loss + NPU BCE
- model.py: forward 增加 dual_window 内联编码，避免 DDP 二次 forward
- model.py: 增加 skip_decoder 参数节省 dual window 计算
- model.py: AEFOutput 扩展 dual_pre_w2 字段
- trainer.py: 主 forward 使用 w1 为默认窗口，内联 dual_window 编码 w2
- trainer.py: 移除独立的 dual window forward（避免 mark-ready + inplace 错误）
- losses.py: change_consistency_loss BCE fallback 到 fp32（NPU bf16 不支持）
- 训练成功恢复 E47，temporal loss 从 9.7→0.38 快速收敛

+++ src/training/losses.py
@@ -547,8 +547,9 @@ def change_consistency_loss(
 
     # BCE Loss: change_score 应该预测图像差异
     # 用 img_diff_norm 作为 target (软标签，不是二值)
+    # ★ NPU bf16 不支持 BCE，fallback 到 fp32
     loss = F.binary_cross_entropy(
-        change_score.squeeze(1), img_diff_norm, reduction='none'
+        change_score.squeeze(1).float(), img_diff_norm.float(), reduction='none'
     )
     loss = (loss * change_mask).sum() / change_mask.sum().clamp(min=1.0)
     return loss
